Fixed: file/overall progress was reset after completion
Previously the file and overall progress was reset to zero after a 1 second delay. This did not allow the user to see that the upload had completed.

Now the progress is left wherever it ends up, it is up to the delegate to reset if so desired.

In addition, adding or removing files resets the overall progress uploaded bytes to zero.