New: filename filter is a string, and setAllowedExtensions was added.
You can now set the filename filter either as a string (-setFilenameFilter:) or as a regular expression (-setFilenameFilterRegex:). In addition, a convenience method, -setAllowedExtensions:, allows you to pass an array of allowed filename extensions (without dots), which is turned into a filename filter. The filter is an inclusion filter.

If a file is filtered out during an add, the new delegate method -fileUpload:didFilterFile: is called.

The demo was updated to add a filter field and an alert when a file is filtered out.