New: the class of the object stored in the queue can be set.
You can now create a subclass of JQueryFileUploadFile and set it as the file class so that you can add extra state and actions to the file. The class is set with JQueryFileUpload -setFileClass:. If the class is not or does not inherit from JQueryFileUploadFile, a warning is generated and nothing happens.

The demo was updated to show how this can be used to provide a custom progress message.